requirements.txt
discord.py
pandas
numpy
matplotlib
pytz
